+++ library/aik099/PHPUnit/APIClient/SauceLabsAPIClient.php
@@ -13,6 +13,9 @@ namespace aik099\PHPUnit\APIClient;
 
 use WebDriver\SauceLabs\SauceRest;
 
+/**
+ * @link https://docs.saucelabs.com/dev/api/jobs/
+ */
 class SauceLabsAPIClient implements IAPIClient
 {
 
@@ -48,14 +51,21 @@ class SauceLabsAPIClient implements IAPIClient
 	/**
 	 * Update status of the test, that was executed in the given session.
 	 *
-	 * @param string  $session_id  Session ID.
-	 * @param boolean $test_status Test status.
+	 * @param string  $session_id          Session ID.
+	 * @param boolean $test_status         Test status.
+	 * @param string  $test_status_message Test status message.
 	 *
 	 * @return array
 	 */
-	public function updateStatus($session_id, $test_status)
+	public function updateStatus($session_id, $test_status, $test_status_message)
 	{
-		return $this->_sauceRest->updateJob($session_id, array('passed' => $test_status));
+		$data = array('passed' => $test_status);
+
+		if ( $test_status_message ) {
+			$data['custom-data'] = array('status_message' => $test_status_message);
+		}
+
+		return $this->_sauceRest->updateJob($session_id, $data);
 	}
 
 }